feature/AP-95  Added status endpoint and legalaidapplications endpoint
As part of these changes i have added
Rspec tests for:
StatusController
LegalaidApplicationController.

The StatusController is for simple health check for the service. I actually used it
to practice my TDD skills in rails.

service should be available at localhost:3000/status

The LegalaidapplicationController is dealing with a post request to create a new application
At the moment this is hard coded to satisfy the test. The next commit will make this bit flexible.
Once i discovered how to mock stuff in ruby and rails.

service should be available at localhost:3000/legalaidapplications method=post