feature/AP-95 : Persistence to sqllite and refactored. Refactored the controlers to be inside a namespace.  so you will find two new directoriies. /api /api/v1. The directories are then namespaced in the routes config and the controllers become part of the Api::V1 modules.
I have connected the create method to the LegalaidApplication active record in order to persist
the changes.

The LegalaidApplicationRecord now has a private method which will generate  the applicaton_ref using
(some guid system for now) before the active record is created. If the logic for this gets
more complex we may want to extract it into a ActiveRecordService and use that to call the persistence layer.

I have also added Fast JsonAPi.  This is a json serialisation api which will allow
us to have control on what gets serialised in our response.  The alternatives for this are
Active Model serialisers (AMS).  I have chosen to use fast Json API as it apears
to have better performance (according to google).
so now in our response we will not get date_created last_updated and the private_key id field being returned.
(We may need to discuss response structure for our service to see if this is adequate otherwise we will have to change strategy.)