Add proceedings seeds
NOTE: This is intented to be a very streamline of the seeding to allow the
developers to work on top of data that is expected to exist in the
database. This was not implemented with a full seed functionality where
extra logging is generated to notify the user of the outcome (hence the
steamline version). People can iterate on top of this.