Trim branch names to keep deployment names < 53 chars
Helm requires deployment names to less than or equal to 53 characters.
This causes failures in UAT environments when branch names are too long.
To fix this, trim the branch name to a suitable length to prevent
this.

Trimming will be done from the end of the name to try to maintain
information about the branch and avoid duplicate names.