Remove validation for 2 digit year of birth
It was previously necessary to have validation to prevent users
entering a two-digit year of birth. This was due to the way the provider
front-end application handled dates of birth as three separate values
which were passed to the api.

Following the move to a single application the way the date of birth is
handled has changed to use a date object and this validation is no
longer needed. Two-digit dates of birth are correctly interpreted by
rails without any validation.