Add timeout to staging/prod helm deploys
Recent circleci deployments to the staging environment have failed due
to an exceeded quota, however feedback from circleci indicates that the
build was successful.

To get more feedback, add a --wait flag to the helm upgrade command so
that if the deployment fails it will timeout and report a failure.

This replicates existing functionality for UAT deployments.