Added if statement to text field partial
This new if statement requires a new local to be passed to the text field partial that determines whether a hint is displayed if a hint is available.
This means that on the address look up page a hint is displayed over postcode put not in the manual address page.
This also means that any text field that has a hint now needs an additional local variable called hint? that is set to ture.
I have added this to the national insurance number text field on the client details page but any other text field that requires a hint will need to be altered.
Alternatively I could change it so that the text field partial assumes the hint is allways to be displayed unless specified with a hint? variable set to false.
That way we won't have to go through the entire app searching for text fields that require a hint.