Retrieve Organisation from postcode API results
There are certain cases where the Ordnance Survey postcode API returns
results which do not have any values which we include in line_one or
line_two of the address. SW1A 1AA is an example of this. This means
that when the address is displayed for selection it is meaningless to
the user.

This commit retrieves the ORGANISATION_NAME from the result and includes
it in the address - it changes the service, form, serializer and model,
and persists it to the addresses table in the database.

It also makes a change to the address model to ensure that if an element
of the address is blank, then the address is displayed properly without
any extra commas.