Improve layout of check your answers page
There are two check your answers sections on this page. One for the
application, which contains proceeding details, and one for the
applicant. These were originally implemented as separate items on the
page, however this meant that if the layout for one changed (due to a
long value being displayed) the appearance of the two sections was
inconsistent.

To fix this, this commit changes the page so that only one check your
answers section exists, and is broken up using <dt> tags as section
headers.