AP-211 Remove created_at from comparison of attributes
The comparison of created_at is failing on CircleCIi:

Citizens::OwnHomeForm#save leaves other attributes on the record unchanged
     Failure/Error: expect(application.send(attr)).to eq(val), "Attr #{attr}: expected #{val}, got #{application.send(attr)}"

       Attr created_at: expected 2018-12-06 12:32:14 UTC, got 2018-12-06 12:32:14 UTC
       Diff:

I think this is due to rounding errors, - the times are certainly equal to the nearest second.
In any case, it's not germaine to the functionality, so this attribute has been removed from the
comparison of before and after update.