Reduce number of UAT instances to 1
UAT deploys are failing because of a lack of resources in the uat
cloud platform environment.

To try and remediate this, reduce the number of instances created from
two to one. This should hopefully reduce resource consumption, while the
resilience offered by a replica instance is not really needed in a
short-lived UAT environment.