AP-222 Provider journey property values page
Created a new PropertyValuesController in the Providers namespace, and referenced the form
from the view in the citizens page.

In that form, the normal t('some_translation_key') was replaced with
controller_t('some_translation_key') which will look up translations based
on the controller path rather than the template location. This enables
us to have different field names depending on which controller it's being used from, e.g.

Enter the value of your home
vs

Enter the value of your client's home