AP-249 Add Save as Draft button to pages in Provider journey
This PR implements a Save and Return feature to the provider journey.
Each page now has a Save as draft button which will sage the updates to
the database and redirec to the Provider applications page, in addition to
the existing Continue button which saves the changes and progresses on to
the next page in the journey.

In the views, forms that previously had a submit button
now render a partial `share/forms/continue_and_draft_buttons` which renders
the two buttons if required (most forms in whch this partial is included are
used for the citizens journey as well, so the partial required a local `show_draft`
to be set to true or false to determine whether or not to render the Save as Draft button.

Pages which were not previously forms, and had the continue
button rendered with the `button_to` now submit a form with just the two
buttons, via the partial `shared/forms/continue_and_draft_button_form`.

In the controllers, the logic of determining which button was used to submit the
form and redirect accordingly is in a Concern SaveAsDraftable, which provides a method
`continue_or_save_draft'.

As well as providing a Save as draft button on all the provider pages, all the
pages are now wired up to continue to the next page in the journey, so the PR incorporates
some logic to provide for that as well.

This PR does NOT address the Back link, which still needs to be wired up correctly
on all the provider pages.