Amend check for pg database
The docker/wait_for_pg script is run when the application is deployed.
It checks that the postgres container has been started before creating
the apply database. This is needed on environments where we are using a
containerised database (ie UAT), however it will now fail where we are
using an RDS instance (ie staging and production). It is not needed in
these environments.

To resolve this, amend the script so that it checks where the
deployment is taking place - if the db username is 'postgres' then it
must be UAT, so run the status check, otherwise don't (staging and
production use a randomly generated db username so this will never flag
a false positive.)