Remove 'Signed in successfully' message
When a user signs in via the portal devise sets a flash message which
is displayed to the user. They don't need to see this.

Remove it by removing the message wording from the
config/locales/devise.en.yml file.

This will allow other flash messages set by devise to still work as
expected (if we need them - if we don't then they can be removed by a
separate change).