Add styling for app-check-your-answers--long
The design system has additional styling for Check Your Answers pages,
the .app-check-your-answers--long class which increases the width of
question field to prevent long questions being split over multiple
lines.

This was removed from our version of the pattern by a previous commit
but is now required.

This change adds back in the styling for the class, and amends the
check_answers_helper and shared partials to allow an extra parameter
(is_long) to be passed so that the --long class can be used from the
existing helper.