Add remaining merits steps to check answers screen
This commit displays values entered on the 'Statement of Case' and
'Prospects of Success' screens on the 'Merits Check Your Answers'
screen.

This commit also tidies up the navigation through all the merits
screens, removes placeholder text, finalises the merits specs and adds
feature test steps for the merits flow.