AP-325 Fix and standardise currency validations
Previously, currency validations were inconsistent and allowed too many decimals.
Have created a new currency validator, which is used on all forms.
In addition, CurrencyCleaner will remove currency signs and commas if the string is a valid value.
If there are errors, the original String that was input by the user is redisplayed back on the form.