Improve text layout on check merits answers screen
On a number of merits questions a provider can enter text explanations.
It is possible that a user will enter multiple paragraphs of text,
especially on the Statement of Case screen. At present when these
responses are displayed on the check your answers screen linebreaks are
ignored which means that multiple paragraphs are displayed as a single
block of text, making it hard to read.

To improve this, use the simple_format method to display text on the
check your answers screen as it was entered by the user.