Ensure unique element ids on CYA pages
When the _one_link_section partial is rendered with nothing declared,
the same element id is given to both the header for the section and the
section that contains the Change link.

To ensure unique element ids, append '_header' to the first id when an
empty answer_hash is provided.