Delete property data if ownership is changed to no
Currently, if a user indicates that they own their property and enters
value/mortgage/shared ownership, then navigates back to the property
questions from the capital check your answers page and selects 'No' for
property ownership, the original values entered for the other property
questions remain in the database.

This commit adds a method to the model to clear those values and a call
from the form to delete them if the value is change to No.