Delete % if shared ownership is changed to no
Currently, if a user indicates shared ownership of their property and
enters a percentage, then navigates back to the shared ownership
question from the capital check your answers page and selects 'No',
the original percentage remains in the database.

This commit adds a call from the form to delete the percentage if
shared ownership is change to No.