Add Identify Types of Outgoing page
This commit adds the necessary code for the page to allow a citizen to
identify their regular outgoings.

The previous and next pages do not yet exist so navigation is not fully
implemented here. For this reason no feature tests have been written.
These will need to be addressed when more of the flow is in place.

- add route
- add controller
- add (minimal) navigation
- add view
- add translations
- add tests