AP-418 Spike CCMS integration
This branch contains a class which generates a SOAP packet to make the
GetReferenceData call to the CCMS test environment on RedCentric.

Currently, the connection fails becaause we haven't yet got a connection to
redcentric, but there is a method to print out the XML which can then be pasted into
a browser window running Remote Desktop and submitting the request that way.
Unfortunately, the paste funciton is extremely buggy, and usually hangs RDP client
if you paste more than three or four lines at a time, so waiting until we have a direct
connection to Redcentric before proceeding further.

This requires the following environment variables to be set in order to generate the request:

- SOAP_CLIENT_USERNAME
- SOAP_CLIENT_PASSWORD_TYPE
- SOAP_CLIENT_PASSWORD
- USER_LOGIN
- USER_ROLE

The values for the test redcentric environment can be found on Rattic under
"Redcentric CCMS test credentials"

ap-418 GetReferenceData produces OK xml

AP-418 remove commented out line