Ap 296 Check Your Answers page is Read Only for applications that are provider_submitted.
I created read_only variable in the controller which is true when the state is provider_submitted. This means that the check your answers page for provider can be read only on the condition that the application is submitted.

Also, When a provider clicks link 'View' on an application that has state provider_submitted, they'll be directed to the read only version of the check your answers page.

Created a helper method for getting the html value of a button.

Updated the provider_step when the provider submits an application and is on the application_confirmation page, so when they view the application from the providers/applicants page, they'll end up on the check your answers and not the application_confirmation page.