Add styling to wrap text on provider profile
The 'role' field on the provider profile page can contain a long list
of roles which does not currently wrap.

This commit adds a css class to force the text to wrap. It also changes
the way the comma-separated list of roles is displayed by adding a
space after each comma, which will prevent the list wrapping mid-word.