AP-392 Merits Assessment Flow
Merits-Assessment-Flow is correct to current requirements

Skip over two pages in the provider merits flow for passported and non-passported applications. This is because these pages are not needed now but could be used in the future for certain proceeding types. The pages are proceedings_before_the_courts and client_received_legal_helps.

Created tests for Merits-Assessment-Flow to check that the flow is correct and updated tests to not include proceedings_before_the_courts and client_received_legal_helps pages.