AP-468 HostingEnvironment class to determine where we are running
Created a HostingEnvironment class to determine which environment we're running in.
Within this class, there is the .env class method to return the environment name in symbol format.
Also, created methods test?, development?, uat?, staging? and live? that return true if it's in said environment.

Lastly, created tests to make sure that the correct environment is returned and that a runtime error is set off if the environment url is unknown.