Bail conditions content on /respondent page
For the question 'Have bail conditions been set?' on the respondent page, the label for its text area should change depending on whether the user selects yes or no as an answer to this question.

Added custom data attributes to store the two labels for Bail conditions.
Using JavaScript in the front-end to change the label depending on the option selected for bail_conditions_set ('yes', 'no').
Update the error text and the error summary list depending on the label.