Add service to poll for CCMS applicant creation
This commit creates the check_applicant_status_service and integrates
it into the submission process.

It introduced the concept of polling, by adding a poll_count to the
ccms_submissions table and a POLL_LIMIT constant. CCMS will be polled
repeatedly until either an applicant has been created or the poll limit
is reached, at which point the process will fail.

The commit also includes small changes to the _requestor and
_response_parser for this process and corrects some test data.