Clicking the View/change declared income links and View/change declared outgoings links on the Means summary page, will take you to the appropriate income_summary, outgoings_summary pages in order to view/change values. Then clicking the save and continue button on income_summary and outgoings_summary will take you back to the means_summary page.
Update means_summaries/show page links for the View/change declared outgoings and View/change income outgoings to route to the outgoings summary and income summary page.
The forward step of the income_summary is means_summary if the application is in the state checking answers.
Add a feature test to show that the routing is correct