Add check to pretty_postcode
pretty_postcode inserts a space into a postcode to put it in the correct
format. It's possible for this method to be called twice, which
incorrectly results in two spaces being inserted.

To fix this, add a new method, pretty_postcode?, to check whether the
postcode is already 'pretty' and use this to control the insert.