Restructure seed file to ensure correct ordering
At the moment all seed files are run in alphabetical order. This hasn't
been a problem up until now, but the legal legal_framework seeds need to
be run in a specific order - proceeding_type_scope_limitation join data
needs to be seeded after proceeding_types and scope_limitations. This
fails using the current approach.

Instead, move the three legal_framework seed files to a separate
directory and add a seed file which will run those three in the correct
order. Other files are not affected by this and will still be run
alphabetically.