Delete unused 'clean_old_releases' code
bin/uat_deploy contains some old code which was intended to delete
old UAT releases to prevent the number building up. It was never
implemented properly and didn't delete anything, but gave the
impression to the user that it did.

Following a recent change to delete old UAT instances when their branch
is delete in GitHub this code is completely redundant, so delete it.