Setup connection to CCMS SOA UAT
We need to connect our service to CCMS SOA UAT environment through
pointing our requests to an AWS API Gateway that points to CCMS SOA UAT
in 6DG as back-end.

In order to be able to do that, we downloaded the WSDLs definitions for
all the endpoints we need and replaced the internal URLs containing
6dg private domains (not accessible from Cloud Platform directly) and
replaced them with the domain/url from the AWS API Gateway.

In this way, the definitions of the WSDLs are still valid, while the
connections go through the API Gateway in order to be able to reach
CCMS SOA.

Add CCMS SOA env vars to Rails configuration

We need these variables to be set up in order to correctly authenticate
our requests to CCMS SOA API.

 By defining them in Rails config we expose them on
"Rails.configuration.x.ccms_soa.***" values.

Setup WSDL files in Rails configuration

This will allow to define different WSDL files per environment in the
environment secrets and access them from the same Rails config values.