Update respondents factory
Update respondents factory, hardcoding value of 'true' and 'false' where relevant.
This is because for example bail conditions details are cleared if the answer is set to 'no'/'false'
This clearing details functionality is something I added into this feature.
This broke the positive tests and thus it uses hardcoded values to allow details to be inputted and saved.