Fix broken start chevron
Upgrading to v3 of the GOVUK design system has caused the > to disappear
from start buttons.

The HTML for it now needs to be added to the button, as per
https://design-system.service.gov.uk/components/button/#start-now-button-example-html

I've created a helper method to format the label, concatenating the
correct wording to the html for the chevron. To enable this I've had to
change a button_to to link_to, as button_to does not allow embedded
HTML in the label.