AP-930 Fix test coverage issues
Hardcoding the 'opponent' and 'other party' data in the CCMS payload
means that several methods are no longer being called, and so are no
longer being executed by specs.

This causes builds to fail due to insufficient code coverage.

This commit adds tests and (temporarily) removes some code to ensure
test coverage is 100%.