AP-810 Fix CCMS Document Upload
Documents (eg statement of case PDFs) aren't currently being uploaded
to CCMS correctly. The current, incorrect, process is:

1. request_case_id
2. create_an_applicant
3. poll_applicant_creation
4. create_case
5. poll_case_creation_result
6. request_document_ids
7. upload_documents

This is incorrect as document ids need to be included in the
create_case payload.

This commit changes the order of CCMS requests to

1. request_case_id
2. create_an_applicant
3. poll_applicant_creation
4. request_document_ids
5. create_case
6. poll_case_creation_result
7. upload_documents

by amending the logic in ccms_submission_state_machine, and changing
various ccms _service classes appropriately.

It also includes ccms_document_ids in the CaseDocs section of the case
creation payload.