Create substantive cases with CCMS integration spec
Amend ccms_integration/ccms_spec.rb so that it injects two cases into
CCMS when it is run - one delegated functions (as it does already) and
one substantive.

- Add an extra context to account for the substantive case. This
involves some refactoring of the existing spec.

- Amend save_request functionality so that it includes the case type in
the filename, to make it easier to see what each payload is.

- Use factories to create proceeding_types and scope_limitations,
instead of doubles.

- Changes to case_add_requestor and ccms_keys to ensure correct values
are being populated in the payload.