AP-916 Refactor CCMS document list to array
The Submission model holds a list of documents to be uploaded to CCMS.
This is a hash of key/value pairs, where the key is the id of the
PDF to be uploaded and the value is the status of the upload.

This worked ok when we had one type of document (statement of case) but
is problematic when there are multiple types (means and merits reports).

There is now a requirement to hold the 'type' of document, as this needs
to be sent to CCMS as part of the upload, and also the ccms_document_id,
which is obtained from CCMS. For statements of case this is held on the
model, however there are no models for means and merits reports so this
report is not suitable.

This commit refactors the submission so that instead of a hash, the
document list is held as an array of hashes, where each hash contains
the id, status, type and ccms_document_id.