AP-903 Fix APPLICATION_FROM_APPLY attribute spec
The APPLICATION_FROM_APPLY attribute is being tested in
case_add_requestor_xml_blocks_spec. It should appear in the payload
twice, once in global_means and once in global_merits. The spec is
testing in global_means both times and so is inaccurate.

This commit fixes this by changing one of the global_means to
global_merits, so both instances are correctly tested.