Add CCMS connectivity test button to Admin screen
To test connectivity to CCMS from a deployed application a way of
triggering a call to CCMS from Apply is required. Rspec tests cannot be
used as they are for local development.

To achieve this, add a button next to each record on the Admin screen.
Clicking this button will use the ReferecnceDataRequestor to call
CCMS. The response from CCMS will be displayed on screen.

Add:
 - controller
 - template to display response
 - spec (although this doesn't run well due to VCR so commented out)

Amend:
 - routes.rb to add new route
 - /admin/legal_aid_applications/_legal_aid_applications.html.erb to
 add button for each record