AP-813 Titleize addresses when returned from API
Put addresses into title case as soon as they are received from the API,
instead of when rendering them on the screen.

This means that they are displayed in a more readable format when a user
views the list of addresses to select from, and also means that the
address they select is persisted to the database in that format.