Fix slack integrations
Following Julien's departure, our sentry integrations to slack are no
longer working. I've readded sentry to the alerts channels but this
has resulted in new webhook urls, which now need to be added to to
the helm values files for UAT, Staging and Production.

This commit replaces the old webhook with the new one in each of those
files.