AP-949 Persist CCMSDocumentIDs to database
Document IDs are not being included in the case creation payload being
sent to CCMS. The CCMSDocumentID attribute is always empty. This data is
needed in order for CCMS to generate a Document Request Task for
caseworkers to process.

This is occurring because CCMS Document IDs are not being persisted to
the database by the ObtainDocumentIdService.

Fix the request_document_ids method so that it calls document.save!
after receiving a document id in order to persist it. This will allow
the AddCaseService to populate the ids correctly.