Add missing staging and production values
The following values are needed for connecting to CCMS SOA:

  awsGatewayApiKey
  clientUsername
  clientPasswordType
  clientPassword
  userLogin
  userRole
  caseServicesWsdl
  clientProxyServiceWsdl
  documentServicesWsdl
  getReferenceDataWsdl

They are held in values files in helm_deploy. The keys have been added
for UAT but are not yet known for staging/production. This omission is
causing failures in CircleCI deployment. This commit adds the keys with
empty strings as values.