AP-956 Fix incorrect CCMS test data
ccms_spec.rb uses factories to create proceeding types and scope
limitations for test applications. The data produced by these factories
by default is not compatible with CCMS and so the cases produced are
not useable. This needs to be fixed by overriding factories in that
spec to force the correct values to be populated.

Additionally, ccms_spec.rb is using incorrect provider details. Some of
the details it currently populates is related to a different firm to
the rest of the provider data.

* create proceeding types and scope limitations for substantive and
delegated functions cases using valid data

* fix incorrect provider details by using data for Desor & Co

* replace an unnecessary ActiveRecord call in case_add_requestor with a
method call to the Limitation model