Save CCMS poll counts to database
When polling for the creation of an applicant or case in CCMS, a
count of polling attempts is kept.

Sidekiq is using a new instance of the polling service each time it is
called, which effectively resets the poll count to 0 each time.

To prevent this, call submission.save each time the poll count is
incremetented.