Add delay to CCMS calls
Sidekiq is calling each step of the CCMS injection process immediately.
This is fine for most steps but causes failures when CCMS has to
process something and a response is not immediate (for example when
creating an applicant or case).

To fix this, add a 5 second delay to each call. This will increase the
overall time taken to create a case, but will prevent individual steps
from failing.