Create Privacy Policy page in the application view footer
Create page for the Privacy Policy statement and link it in the footer of the application view

Create page for the Privacy Policy statement and link it in the footer of the application view

Create controller, view, translation and tests for the privacy_policies page

Link the privacy policy in the footer layout of the application view

Update list_from_translation_path method to handle translation where the journey is unknown to prefix the translation path with an empty string

To note:
There are embedded hyperlink texts in the privacy policy statement. I embedded the hyperlink in the translation file and appended _html at the end of the translation key. This meant that I was not able to iterate through a translation list if it had embedded links in it.