AP-160 Add Production Benefit Checker creds
Up until now production has been configured to use the staging benefit
checker instance. Now we are close to going live it needs to be changed
to use the production benefit checker.

This commit changes the values-production helm file to use the correct
clientOrgId and clientUserId keys and point to the production benefit
checker url.