AP-852 Split accounts checkbox into current/savings
The first checkbox on the savings_and_investment screen needs to be
split into two checkboxes to improve user experience.

This commit

* amends the savings_amounts table to hold two values, not just one
* changes label, hint and error text in translation files for both
  provider and citizen screens
* amends the CFE service to ensure all capital values are included
* fixes tests that break as a result of those changes

It also makes a number of other changes to the wordings on the
savings_and_investment screen to bring it line with the prototype (for
example, changing 'Cash savings' to 'Money not in a bank account').