AP-1042 Remove hardcoded provider reference
All cases injected to CCMS are including a ProviderCaseReferenceNumber
of 'PC4'. This will be confusing for solicitors.

Apply doesn't currently ask the provider for their reference as part of
the application process.

For now the agreed fix is to remove the ProviderCaseReferenceNumber
from the case creation payload. If solicitor feedback from beta testing
indicates they need it, then it can be added back along with a change to
the Apply journey to allow a solicitor to enter it.