AP-914 Add CCMS case reference to means report
The CCMS case reference needs to be displayed on the generated means
report PDF so that caseworkers are able to track applications.

There is already a pattern for doing this, as it has been implemented
previously on the merits report.

This commit adds code to the means_reports_controller to get the case
reference from CCMS if it does not already exist, and amends tests
appropriately.

The shared partial to display the reference has already been amended in
the commit that added the case reference to the merits report, so no
change is necessary.