AP-914 Add capital results to means report
Caseworkers need to see the outcome of the capital assessment to
properly assess the case.

This commit:
* Adds a partial containing the total capital, capital thresholds and
capital contribution amount and include it in the
means_reports/show.html.erb view.

* Adds wordings to translation files, including values for the lower and
upper capital thresholds which aren't defined anywhere else in the
application.

* Adds a trait to create a cfe_result for a legal_aid_application, and
uses it to ensure a cfe_result always exists before a means_report is
created. It also uses that trait in various specs, replacing let blocks.