AP-1020 Store provider user_login_id
A provider user's user_login_id is currently mocked by the Provider
model, which means that all submissions to CCMS contain the same id.
This is incorrect.

Add a column to the providers table to hold this data, and extract it
from the provider details api payload when creating a provider.