AP-1044 Correct benefit check result
Attribute LAR_INFER_B_1WP1_36A is being populated with
application.benefit_check_result.

benefit_check_result is an object, so this appears in the xml payloads
like: #&lt;BenefitCheckResult:0x000055f9a5681a38&gt;

This is a boolean attribute, so this commit changes it to populate true
if application.benefit_check_result.result is yes, and false otherwise.