AP-989 Add Sentry alerting for postcode lookup
We want to be alerted when calls to the Ordnance Survey Places API fail.

This commit adds Sentry calls to the AddressLookupService in two places,
so that alerts will occur if the service is unavailable and if the
service is available but is unsuccessful in returning a response (for
example if authentication fails).