AP-1019 Prevent extra document id request
Investigating various issues with uploading documents to CCMS, I noticed
that we always make an extra unnecessary document id request (ie if
there are two documents, three requests are made; three documents, four
requests).

This is occurring because the memoized response method is being called
when the history is written, which is causing an additional request to
be made.

This commit fixes that by changing response to an instance variable,
which is assigned to in the request_document_ids and simply read from
when creating the history record.

This change highlighted the fact that there is some unnecessary logic
present in the ObtainDocumentIdService class to create a history
record when no documents exist. This is also done in the AddCaseService
class, so this commit removes that line and moves some of the tests to
the AddCaseServiceSpec instead.