AP-1033 Omit unnecessary CCMS attributes
Four attributes are currently being included in CCMS case creation
payloads that are not needed. They are marked as OMIT on the
CCMS payload state of play spreadsheet.

This commit set the generate_block? flag to false for all four
attributes in the ccms_keys and adds them to the omit list in the
case_add_requestor_xml_blocks_spec.