Restrict logins to authorised beta users in production
Update Authorizable module that uses Pundit to authorise access to only those users that are whitelisted whilst in production. If the user is not authorised, redirect them to an access_denied page

Store whitelisted users/providers in encrypted file

Call whitelisted users in provider model to return a boolean on whether the provider is whitelisted or not.

Test provider model and the redirection to an error page for non-whitelisted providers