AP-1059 CCMS: Use correct warning_letter_sent values
The WARNING_LETTER_SENT attribute is being hardcoded as false, and
INJ_REASON_NO_WARNING_LETTER is being hardcoded as a string.

This commit changes them so it they are populated with correct values
from the database.

* WARNING_LETTER_SENT is populated with respondent.warning_letter_sent

* INJ_REASON_NO_WARNING_LETTER is populated with
respondent.warning_letter_sent_details if no warning letter has been
sent, and is omitted if a letter has been sent