AP-1016 Fix CCMS delegated functions flag
The DelegatedFunctionsApply attribute is not being correctly populated.
It is being set to false on DF cases when it should be true.

This is occurring because the attribute is populated with
!limitation.substantive, I think on the assumption that if it’s not a
substantive limitation then it must be delegated functions. However this
is not the case, a limitation can apply to both substantive and
delegated functions cases.

This commit changes that line so that DelegatedFunctionsApply is
populated with limitation.delegated_functions.