AP-1057 Amend hard coded text in CCMS payloads
A number of attributes in CCMS payloads are hardcoded with a string
identifying it as being an Apply application.

This text is being exposed to providers when amendments are carried out,
which is not desirable (it is intended just for caseworkers).

Replace the string with a '.' to remove this issue.