AP-1084 change COST_LIMIT_CHANGED_FLAG to text
The COST_LIMIT_CHANGED_FLAG should be a text value, not boolean (even
though it can only ever contain true/false...)

This commit changes the ResponseType in the ccms_keys.yml and fixes the
spec. It also removes an unnecessary `COST_LIMIT_CHANGED_FLAG` from the
list of `false` attributes - this attribute has its own spec so doesn't
need to be included there.