AP-1032 Fix Admin Delete All functionality
The button in the Admin screen to Delete All cases does not currently
work. It fails due to FK violations in various relations.

This commit does two things:

* It fixes the violations by adding 'dependent: :destroy' to the
CCMS::Submission model for SubmissionHistory and SubmissionDocument
relations, and 'on_delete: :cascade' to three foreign keys. This will
ensure that all data can be successfully deleted.

* Updates the 'with a lot of associations' test in
admin/legal_aid_applications_spec. This test was previously using an
application created with the with_everything trait. This trait has not
been updated as extra associations have been added to the app, and so
the test was passing even though the functionality did not work. Adding
extra  associations to this trait now is problematic as it causes many
other tests to fail and increases the execution time of the test suite
dramatically.

Instead this commit creates a new trait :at_assessment_submitted, which
represents an application at the end of the process, and uses it in the
'with a lot of associations' test. This includes all present
associations. It will need to be kept up-do-date as more are added
otherwise this will become a problem again in future.