AP-1037 CCMS Populate REQUESTED_SCOPE correctly
REQUESTED_SCOPE is being hardcoded in two different ways - MULTIPLE in
one place and a string identifying it as an Apply case in another.

Both are incorrect - it should be populated as MULTIPLE if more than
one scope limitation exists for the proceeding, or the scope limitation
code if only one exists.

This commit fixes the problem by adding a method to the
attribute_value_generator to generate the correct result and calling it
from the ccms_keys file.