AP-1025 Update Google OAuth2 secrets
Google OAuth2 is being used to authenticate access to the Admin screens
in Apply. The credentials that this was set up with belong to a
developer who has since left the team.

A new dedicated Apply for Legal Aid google project has been set up and
new credentials created. This commit updates the GOOGLE_CLIENT_ID and
GOOGLE_CLIENT_SECRET environment variables to use the correct
credentials.